Remove create_demo_config from genesis_config.rs
