+++ examples/demo-stf/src/genesis_config.rs
@@ -15,23 +15,6 @@ pub const LOCKED_AMOUNT: u64 = 200;
 pub const TEST_SEQ_PUB_KEY_STR: &str = "seq_pub_key";
 pub const TEST_TOKEN_NAME: &str = "sov-test-token";
 
-///
-/// * `value_setter_admin_private_key` - Private key for the ValueSetter module admin.
-/// * `election_admin_private_key` - Private key for the Election module admin.
-pub fn create_demo_config(
-    initial_sequencer_balance: u64,
-    value_setter_admin_private_key: &DefaultPrivateKey,
-    election_admin_private_key: &DefaultPrivateKey,
-) -> GenesisConfig<DefaultContext> {
-    create_demo_genesis_config::<DefaultContext>(
-        initial_sequencer_balance,
-        generate_address::<DefaultContext>(TEST_SEQ_PUB_KEY_STR),
-        TEST_SEQUENCER_DA_ADDRESS.to_vec(),
-        value_setter_admin_private_key,
-        election_admin_private_key,
-    )
-}
-
 pub fn create_demo_genesis_config<C: Context>(
     initial_sequencer_balance: u64,
     sequencer_address: C::Address,
