+++ examples/demo-stf/src/tests/mod.rs
@@ -1,11 +1,19 @@
 use borsh::BorshSerialize;
 use sov_app_template::{AppTemplate, Batch};
-use sov_modules_api::{default_context::DefaultContext, Address};
+use sov_modules_api::{
+    default_context::DefaultContext, default_signature::private_key::DefaultPrivateKey, Address,
+};
 use sov_state::ProverStorage;
 use std::path::Path;
 
 use crate::{
-    app::DemoApp, runtime::Runtime, tx_hooks_impl::DemoAppTxHooks,
+    app::DemoApp,
+    genesis_config::{
+        create_demo_genesis_config, generate_address, TEST_SEQUENCER_DA_ADDRESS,
+        TEST_SEQ_PUB_KEY_STR,
+    },
+    runtime::{GenesisConfig, Runtime},
+    tx_hooks_impl::DemoAppTxHooks,
     tx_verifier_impl::DemoAppTxVerifier,
 };
 
@@ -30,4 +38,18 @@ pub fn create_new_demo(
     let tx_hooks = DemoAppTxHooks::new();
     let tx_verifier = DemoAppTxVerifier::new();
     AppTemplate::new(storage, runtime, tx_verifier, tx_hooks)
+}
+
+pub fn create_demo_config(
+    initial_sequencer_balance: u64,
+    value_setter_admin_private_key: &DefaultPrivateKey,
+    election_admin_private_key: &DefaultPrivateKey,
+) -> GenesisConfig<DefaultContext> {
+    create_demo_genesis_config::<DefaultContext>(
+        initial_sequencer_balance,
+        generate_address::<DefaultContext>(TEST_SEQ_PUB_KEY_STR),
+        TEST_SEQUENCER_DA_ADDRESS.to_vec(),
+        value_setter_admin_private_key,
+        election_admin_private_key,
+    )
 }
